# Conflicts:
#	nhs/src/main/java/uk/ac/cf/spring/nhs/Diary/Controller/DiaryController.java
#	nhs/src/main/java/uk/ac/cf/spring/nhs/Diary/Entity/DiaryEntry.java
#	nhs/src/main/java/uk/ac/cf/spring/nhs/Diary/Service/DiaryEntryService.java
#	nhs/src/main/resources/static/css/common/defaultLayout.css
#	nhs/src/main/resources/templates/diary/diary.html

+++ nhs/src/main/java/uk/ac/cf/spring/nhs/Diary/Repository/DiaryEntryRepository.java
@@ -13,7 +13,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface DiaryEntryRepository extends JpaRepository<DiaryEntry, Integer> {
+public interface DiaryEntryRepository extends JpaRepository<DiaryEntry, Long> {
     List<DiaryEntry> findByUserId(long userId, Sort sort);
 
     List<DiaryEntry> findByUserIdAndDateBetween(Long userId, LocalDate startDate, LocalDate endDate);
